requirements-dev.txt
ipython==7.2.0
awsebcli==3.14.9
django-extensions==2.1.4
djangorestframework==3.9.1